requirements.txt
# Core Django
django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-ratelimit==4.1.0
django-extensions==4.1

# Security & Authentication
django-two-factor-auth[phonenumbers]==1.18.1
django-otp==1.6.1
django-phonenumber-field==8.3.0
cryptography==46.0.3
bcrypt==5.0.0

# Configuration
python-decouple==3.8

# Image Processing
pillow==12.0.0

# Payment Processing
paystack-sdk==1.0.1

# QR Code Generation
qrcode==8.2

# Phone Number Validation
phonenumbers==8.13.55

# Database (for production)
# psycopg2-binary==2.9.7  # PostgreSQL
# mysqlclient==2.2.0      # MySQL

# Development Tools
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Testing
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.38.0

# Additional Utilities
requests==2.31.0
python-dateutil==2.9.0
six==1.17.0
urllib3==2.5.0